Add Naturals, Naturals0, and UniversalSet to the 'from sympy import *' namespace
Also added tests that all singleton classes are imported (with exceptions for
rational numbers and NegativeInfinity, which can be accessed with -oo). The
existing singleton test was moved to the new test_singleton.py file.
