+++ sympy/solvers/__init__.py
@@ -33,6 +33,7 @@ from .decompogen import decompogen
 
 from .solveset import solveset, linsolve, linear_eq_to_matrix, nonlinsolve, substitution
 
+# This is here instead of sympy/sets/__init__.py to avoid circular import issues
 from ..core.singleton import S
 Complexes = S.Complexes
 del S
